+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 - Passage en UTF-8 et adaptation du code
 - Modification de la licence de Cecill version 2 vers AGPL version 3
 
+## Suppression
+- les constantes suivantes ont été supprimées, car obsolète : OPENSIGN_WSDL, OPENSIGN_CA, OPENSIGN_CRT, 
+OPENSIGN_TIMEOUT, ACTES_TYPE_PJ_IS_MANDATORY, DB_CLIENT_ENCODING, PHP_UNIT_AUTOLOADER, MODE, ETAT_CIVIL_FILES_UPLOAD_ROOT,
+IMAP_LOGIN, IMAP_PASS
+- les constantes suivantes ont été supprimées, car elles étaient dépréciées : VERIFICATION_SIREN
+- les constantes suivantes ont été supprimées, car elles ne sont plus nécessaires du fait du passage en docker : 
+XML_STARLET_PATH, MODE_BEANSTALKD, MODE_REDIS
+
 # 4.3.18 - 2022-06-13
 
 ## Correction